catalog: extract pricing.py — pure pricing math off the god-module (#38 step 1)
Audit finding #1 (high): catalog.py (3,374 lines, 24 importers) mixes pricing
math, data literals, import-time ingestion, and routing on the money path.
Step 1 of the safe split: move the 19 pricing symbols (markup 1.10 + floor,
per-token cost, cache-token pricing, PriceTier + select_price_tier,
provider-manifest price parsing) into pricing.py — pure functions of the money
primitives with ZERO dependency on the catalog data (verified: no PROVIDERS/
MODELS refs). catalog.py re-imports them for backward-compat, so all 24
importers are unaffected. A pricing change is now reviewable in isolation.

Facade verified: catalog.PriceTier/select_price_tier/cache_token_prices_
microdollars still resolve; MODELS/MODEL_ENDPOINTS still build at import
(270 models / 701 endpoints). Full suite 1173 green, mypy clean. Data/ingest/
routing extractions follow as separate PRs.
